Update Longest Increasing Subsequence with reconstruction.cpp

+++ Longest Increasing Subsequence with reconstruction.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
-unsigned int n, i, j, k, l, r, mid, pos;
-long long int v[100000], dp[100000], aux[100000], ans[100000];
+uint32_t n, i, j, k, l, r, mid, pos;
+int64_t v[100000], dp[100000], aux[100000], ans[100000];
 int main() {
     std :: cin >> n;
     for (i = 1; i <= n; ++ i)
